# swaggeroutput.json
      {
        "openapi": "3.0.4",
        "info": {
          "title": "Home Energy Api V1",
          "version": "v1"
        },
        "paths": {
          "/v1/authentication/register": {
            "post": {
              "tags": [
                "AuthenticationV"
              ],
              "requestBody": {
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/UserDtoV1"
                    }
                  },
                  "text/json": {
                    "schema": {
                      "$ref": "#/components/schemas/UserDtoV1"
                    }
                  },
                  "application/*+json": {
                    "schema": {
                      "$ref": "#/components/schemas/UserDtoV1"
                    }
                  }
                }
              },
              "responses": {
                "200": {
                  "description": "OK"
                }
              }
            }
          },
          "/v1/authentication/token": {
            "post": {
              "tags": [
                "AuthenticationV"
              ],
              "requestBody": {
                "content": {
                  "application/json": {
                    "schema": {
                      "$ref": "#/components/schemas/UserDtoV1"
                    }
                  },
                  "text/json": {
                    "schema": {
                      "$ref": "#/components/schemas/UserDtoV1"
                    }
                  },
                  "application/*+json": {
                    "schema": {
                      "$ref": "#/components/schemas/UserDtoV1"
                    }
                  }
                }
              },
              "responses": {
                "200": {
                  "description": "OK"
                }
              }
            }
          },
...
        },
        "components": {
          "schemas": {
...
            "UserDtoV1": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "nullable": true
                },
                "password": {
                  "type": "string",
                  "nullable": true
                },
                "role": {
                  "type": "string",
                  "nullable": true
                },
                "homeStreetAddress": {
                  "type": "string",
                  "nullable": true
                }
              },
              "additionalProperties": false
            },
...
          },
          "securitySchemes": {
            "ApiKey": {
              "type": "apiKey",
              "description": "Custom header for API key",
              "name": "X-Api-Key",
              "in": "header"
            },
            "Bearer": {
              "type": "http",
              "description": "JWT Authorization header. Example: 'Bearer {your token}'",
              "scheme": "Bearer",
              "bearerFormat": "JWT"
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "Bearer": [ ]
          }
        ]
      }